refactor: Make the GBTS layer index 16 bit and drop the Options constructor
The node storage already stored the layer index in 16 bits, so make
GbtsLayerIndex itself std::uint16_t, remove the narrowing cast and the
comments that said the index was narrowed.

GraphBasedTrackSeeder::Options is a plain aggregate; the bFieldInZ
constructor is replaced by designated initialisation at the two call
sites.

+++ Core/include/Acts/Seeding/GbtsLayerDescription.hpp
@@ -18,7 +18,7 @@ using GbtsExperimentLayerId = std::uint32_t;
 
 /// Position of a layer within one GbtsGeometry, dense from zero. Ask
 /// GbtsGeometry::layerIndex for it rather than deriving it.
-using GbtsLayerIndex = std::uint32_t;
+using GbtsLayerIndex = std::uint16_t;
 
 /// Where a GBTS layer sits, which fixes its coordinate convention.
 enum class GbtsLayerType : std::uint8_t { Barrel = 0, Endcap = 1 };
